+++ postgres-types/CHANGELOG.md
@@ -14,7 +14,7 @@
 ### Added
 
 * Added `ToSql` and `FromSql` implementations for `Box<str>`.
-* Added `BorrowToSql` implementations for `Box<dyn ToSql + Sync>` and `Box<dyn ToSql + Sync + Send>`.
+* Added `BorrowToSqlChecked` implementations for `Box<dyn ToSqlChecked + Sync>` and `Box<dyn ToSqlChecked + Sync + Send>`.
 * Added support for `cidr` 0.2 via the `with-cidr-02` feature.
 * Added conversions between the `LTREE`, `LQUERY` and `LTXTQUERY` types and Rust strings.
 * Added support for `uuid` 1.0 via the `with-uuid-1` feature.